Fix aice: rename intentResolvers
To be consistent with Class and other directories
path : ./src/intentsResolver => ./src/intentResolvers
file: IntentsResolver => IntentResolver
file: IntentsResolverManager => IntentResolverManager
file: RemoveIntentsResolver => RemoveIntentResolver
file: SimpleIntentsResolver => SimpleIntentResolver
