+++ packages/aice/src/intentResolvers/index.js
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) 2015-present, CWB SAS
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import SimpleIntentResolver from './SimpleIntentResolver';
+import IntentsResolverManager from './IntentResolverManager';
+import IntentResolver from './IntentResolver';
+import RemoteIntentResolver from './RemoteIntentResolver';
+
+export { IntentsResolverManager, IntentResolver, SimpleIntentResolver, RemoteIntentResolver };
