--- packages/aice/src/intentsResolver/index.js
@@ -1,13 +1,0 @@
-/**
- * Copyright (c) 2015-present, CWB SAS
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import SimpleIntentsResolver from './SimpleIntentsResolver';
-import IntentsResolverManager from './IntentsResolverManager';
-import IntentsResolver from './IntentsResolver';
-import RemoteIntentsResolver from './RemoteIntentsResolver';
-
-export { IntentsResolverManager, IntentsResolver, SimpleIntentsResolver, RemoteIntentsResolver };
