+++ packages/aice/test/inputRevolver/inputResolver.test.js
@@ -5,35 +5,35 @@
  * LICENSE file in the root directory of this source tree.
  */
 import chai from 'chai';
-import { IntentsResolver } from '../../src/intentsResolver';
+import { IntentResolver } from '../../src/intentResolvers';
 
 const { expect } = chai;
 
 describe('IntentsResolver', () => {
   it('Should throw error if no settings with name provided', () => {
-    expect(() => new IntentsResolver()).to.throw('Invalid IntentsResolver constructor - Missing name');
+    expect(() => new IntentResolver()).to.throw('Invalid IntentResolverconstructor - Missing name');
   });
 
   it('Should throw error if no name provided', () => {
-    expect(() => new IntentsResolver({})).to.throw('Invalid IntentsResolver constructor - Missing name');
+    expect(() => new IntentResolver({})).to.throw('Invalid IntentResolverconstructor - Missing name');
   });
 
   it('Should train model - Empty case', async () => {
-    const resolver = new IntentsResolver({ name: 'test-resolver' });
+    const resolver = new IntentResolver({ name: 'test-resolver' });
     await resolver.train();
 
     expect(resolver.inputs).to.eql([]);
   });
 
   it('Should train model', async () => {
-    const resolver = new IntentsResolver({ name: 'test-resolver' });
+    const resolver = new IntentResolver({ name: 'test-resolver' });
     await resolver.train([1]);
 
     expect(resolver.inputs).to.eql([1]);
   });
 
   it('Should execute with filter using LANG', async () => {
-    const resolver = new IntentsResolver({ name: 'test-resolver' });
+    const resolver = new IntentResolver({ name: 'test-resolver' });
     await resolver.train([{ lang: 'fr', topic: '*', id: 1 }, { lang: 'en', topic: '*', id: 2 }]);
 
     const result = await resolver.execute('fr', '', {});
@@ -49,7 +49,7 @@ describe('IntentsResolver', () => {
       return [inputs[1]];
     };
     const cbExecute = () => [{ id: 1, lang: 'us' }];
-    const resolver = new IntentsResolver({ name: 'test-resolver', cbTrain, cbExecute });
+    const resolver = new IntentResolver({ name: 'test-resolver', cbTrain, cbExecute });
     await resolver.train([{ lang: 'fr', topic: '*', id: 1 }, { lang: 'en', topic: '*', id: 2 }]);
     expect(resolver.inputs.length).to.equal(1);
     const result = await resolver.execute('en', '', {});
@@ -65,7 +65,7 @@ describe('IntentsResolver', () => {
       return [inputs[1]];
     };
     const cbEvaluate = () => [{ id: 1, lang: 'us' }];
-    const resolver = new IntentsResolver({ name: 'test-resolver', cbTrain, cbEvaluate });
+    const resolver = new IntentResolver({ name: 'test-resolver', cbTrain, cbEvaluate });
     await resolver.train([{ lang: 'fr', topic: '*', id: 1 }, { lang: 'en', topic: '*', id: 2 }]);
     expect(resolver.inputs.length).to.equal(1);
     const result = await resolver.evaluate('en', '', {});
